fixes after rename
